# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {
            "name": "Tengku Romansyah",
            "url": "https://linkedin.com/in/tengkuroman",
            "email": "tengku.romansyah@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "paths": {
        "/auth/order/v1/order/delete/{order_detail_id}": {
            "delete": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Delete user's order. A user only can delete their own order.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order Service"
                ],
                "summary": "Delete user's order.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Param required.",
                        "name": "order_detail_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/auth/order/v1/order/payment/checkout/{order_detail_id}": {
            "patch": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Pay the selected order. A user can only pay their own order.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order Service"
                ],
                "summary": "Pay the order.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Param required.",
                        "name": "order_detail_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/auth/order/v1/order/payment/{order_detail_id}/{payment_provider_id}": {
            "patch": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Select payment merchant after checkout (order created). A user can only pay their own order.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order Service"
                ],
                "summary": "Select payment provider.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Param required.",
                        "name": "order_detail_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Param required.",
                        "name": "payment_provider_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/auth/order/v1/orders": {
            "get": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Get all user's order. Order retrieved only that made by logged user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order Service"
                ],
                "summary": "Get all user's order.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "BearerToken": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}